Update AddProductAction.java
added functionality to the Add Product type button. It now asks the user which product type to add(Grocery, Electronics), then asks for user input of the name, and skuCode(if electronics). Then it shows a dialog to the user that the product has been added to the stock list and that it is also saved to the corresponding text files. So when the user selects button 1 again it will now show the updated stock list from button 2.

Also tweaked InventoryManagementSystemGUI.java to make sure that the add button works