Update BuyProductAction.java
added appropriate error handling for the buy product, makes sure that the user input is not empty, makes sure that the quantity is also valid, it also asks the user again until a valid quantity is entered. also did modifications to the gui by removing the numbers on the buttons to make it more tidy.