New AddProductAction.java
reworked the add product button. it now adds the user input to the InventoryDB. user is presented with a combobox for choosing grocery or electronics. then user inputs the product id which must be unique. if it is then it is added to the InventoryDB. also all the fields must be field otherwise it wont be added to the DB.