New RemoveProductAction.java
updated the remove button function to now interact with the DB instead of the text files (using the old inventory manager). checks if the product exists in the DB based on ID or name, if it does exist then product will be removed from the DB if not then the user will be notified that the product does not exist in the DB. also it checks if the user cancels or the user enters an empty value.